+++ Tests/Unit/DependencyInjection/MassiveSearchExtensionTest.php
@@ -29,6 +29,39 @@ class MassiveSearchExtensionTest extends AbstractExtensionTestCase
         $this->load();
     }
 
+    public function provideLocalizationStrategy()
+    {
+        return array(
+            array(
+                array(
+                ),
+                'noop',
+                array(
+                    'localization_strategy' => 'noop',
+                ),
+                'noop',
+                array(
+                    'localization_strategy' => 'index',
+                ),
+                'index',
+            ),
+        );
+    }
+
+    /**
+     * @dataProvider provideLocalizationStrategy
+     */
+    public function testLocalizationStrategy($config, $expectedStrategy)
+    {
+        $this->load($config);
+
+        $aliasTarget = $this->container->getAlias('massive_search.localization_strategy');
+        $this->assertEquals(
+            'massive_search.localization_strategy.' . $expectedStrategy,
+            $aliasTarget
+        );
+    }
+
     public function provideAdapterConfig()
     {
         return array(